% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivreg.R
\name{fitivreg}
\alias{fitivreg}
\title{2SLS Regression}
\usage{
fitivreg(y, X, Z, wt = NULL, singular.ok = FALSE, qr = TRUE)
}
\arguments{
\item{y}{The response variable, a vector.}

\item{X}{The explanatory-variables model matrix.}

\item{Z}{The instrumental-variables model matrix.}

\item{wt}{An optional weight vector.}

\item{singular.ok}{Is perfect collinearity tolerable in either stage regression?}

\item{qr}{Include the QR decomposition in the returned object.}
}
\value{
A list with the following components:
\describe{
  \item{\code{n}}{number of cases}
  \item{\code{p}}{number of columns of \code{X}}
  \item{\code{q}}{number of columns of \code{Z}}
  \item{\code{qr}}{QR decomposition for stage-2 regression}
  \item{\code{rank}}{rank of model matrix for stage-2 regression}
  \item{\code{qr.1}}{QR decomposition for stage-1 regression}
  \item{\code{rank.1}}{rank of \code{Z}}
  \item{\code{coefficients}}{estimated regression coefficients}
  \item{\code{coefficients.1}}{coefficient matrix from stage-1 regression}
  \item{\code{vcov}}{covariance matrix of estimated coefficients}
  \item{\code{df.residual}}{residual degrees of freedom for stage-2 regression}
  \item{\code{sigma}}{residual standard deviation}
  \item{\code{residuals}}{response residuals}
  \item{\code{fitted}}{model fitted values}
  \item{\code{residuals.1}}{matrix of stage-1 residuals}
  \item{\code{fitted.1}}{matrix of stage-1 fitted values}
  \item{\code{residuals.2}}{stage-2 residuals}
  \item{\code{fitted.2}}{stage-2 fitted values}
  \item{\code{weights}}{model weights or \code{NULL}}
}
}
\description{
The work-horse 2SLS function, not normally meant to be called directly
but rather via \code{\link{ivreg}}.
}
\examples{
m <- with(Kmenta, fitivreg(Q, cbind(1, P, D), cbind(1, D, F, A)))
names(m)
m$coefficients # estimates
sqrt(diag(m$vcov)) # standard errors
}
\seealso{
\code{\link{ivreg}}
}
\author{
John Fox \email{jfox@mcmaster.ca}
}
