% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivreg2.R
\name{ivreg2}
\alias{ivreg2}
\title{2SLS Least Squares Estimation of a Linear Model}
\usage{
ivreg2(
  formula,
  instruments = rhs(formula),
  data,
  subset,
  weights,
  na.action = getOption("na.action"),
  contrasts = NULL,
  singular.ok = FALSE,
  model = TRUE,
  x = TRUE,
  y = TRUE,
  qr = TRUE,
  ...
)
}
\arguments{
\item{formula}{A model \link[stats]{formula}, as for \code{\link[stats]{lm}},
specifying the response variable and the right-hand side of the model.}

\item{instruments}{A one-sided formula, specifying the instrumental variables.}

\item{data}{An optional data frame, list, or environment containing the variables in
the model formula and instrumental-variables formula.}

\item{subset}{See \code{\link[stats]{lm}}.}

\item{weights}{An optional vector of inverse-variance weights; if absent the weights
are set to 1 for all cases.}

\item{na.action}{See \code{\link[stats]{lm}}.}

\item{contrasts}{See \code{\link[stats]{lm}}.}

\item{singular.ok}{If \code{FALSE} (the default) perfect collinearity in either stage of
the 2SLS regression produces an error.}

\item{model}{Include the model frame in the returned object.}

\item{x}{Include the model matrices for the two stages in the returned object.}

\item{y}{Include the response vector in the returned object.}

\item{qr}{Include the QR decompositions for the two stages in the returned object.}

\item{...}{Not used.}
}
\value{
An object of class \code{"ivreg2"}, with the following elements in addition to
those returned by \code{\link{fitivreg2}}:
\describe{
  \item{\code{response.name}}{a character string with the name of the response
  variable or possibly the character representation of an expression evaluating
  to the response.}
  \item{\code{formula}}{the model formula.}
  \item{\code{intruments}}{the one-sided formula specifying the instrumental variables.}
  \item{\code{model.matrix}}{the model matrix corresponding to the right-hand side of
  the model formula.}
  \item{\code{y}}{the response-variable vector.}
  \item{\code{model.matrix.instruments}}{the model matrix for the first-stage regression.}
  \item{\code{wt.var}}{the name of the weight variable (for a weighted fit).}
  \item{\code{na.action}}{the na.action argument.}
  \item{\code{call}}{the function call.}
  \item{\code{contasts}}{contrasts for the model matrix.}
  \item{\code{contrasts.instruments}}{contrasts for the instrumental-variables
  model matrix.}
  \item{\code{xlevels}}{factor levels (if any) for the model matrix.}
  \item{\code{xlevels.instruments}}{factor levels (if any) for the
  instrumental variables model matrix.}
  \item{\code{terms}}{the terms object for the model matrix.}
  \item{\code{terms.instruments}}{the terms object for the instrumental
  variables model matrix.}
  \item{\code{model}}{the model frame.}
  }
}
\description{
Provides a formula-based interface for 2SLS estimation of
a linear model. Computations are done by \code{\link{fitivreg2}}. The returned object has
the necessary information for computing a variety of
\link[=ivregDiagnostics]{regression diagnostics}.
}
\examples{
summary(ivreg2(Q ~ P + D, ~ D + F + A, data=Kmenta))     # demand equation
summary(ivreg2(Q ~ P + F + A, ~ D + F + A, data=Kmenta)) # supply equation

}
\seealso{
\code{\link{fitivreg2}}, \code{\link[stats]{lm}}, \code{\link[stats]{formula}}, \code{\link{2SLS_Methods}},
\code{\link{ivregDiagnostics}}
}
\author{
John Fox \email{jfox@mcmaster.ca}
}
