% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivregHexsticker.R
\name{ivregHexsticker}
\alias{ivregHexsticker}
\title{View the Official Hex Sticker for the ivreg Package}
\usage{
ivregHexsticker()
}
\description{
Open the official hex sticker for the \pkg{ivreg} package in your browser.
}
\examples{
\dontrun{
ivregHexsticker()
}
}
