% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivregMethodsAER.R
\name{vcov.ivreg}
\alias{vcov.ivreg}
\alias{ivregMethods}
\alias{bread.ivreg}
\alias{estfun.ivreg}
\alias{hatvalues.ivreg}
\alias{terms.ivreg}
\alias{model.matrix.ivreg}
\alias{predict.ivreg}
\alias{print.ivreg}
\alias{summary.ivreg}
\alias{print.summary.ivreg}
\alias{anova.ivreg}
\alias{update.ivreg}
\title{Methods for \code{"ivreg"} Objects}
\usage{
\method{vcov}{ivreg}(object, ...)

\method{bread}{ivreg}(x, ...)

\method{estfun}{ivreg}(x, ...)

\method{hatvalues}{ivreg}(model, ...)

\method{terms}{ivreg}(x, component = c("regressors", "instruments"), ...)

\method{model.matrix}{ivreg}(object, component = c("projected",
  "regressors", "instruments"), ...)

\method{predict}{ivreg}(object, newdata, na.action = na.pass, ...)

\method{print}{ivreg}(x, digits = max(3, getOption("digits") - 3), ...)

\method{summary}{ivreg}(object, vcov. = NULL, df = NULL,
  diagnostics = FALSE, ...)

\method{print}{summary.ivreg}(x, digits = max(3, getOption("digits") -
  3), signif.stars = getOption("show.signif.stars"), ...)

\method{anova}{ivreg}(object, object2, test = "F", vcov. = NULL, ...)

\method{update}{ivreg}(object, formula., ..., evaluate = TRUE)
}
\arguments{
\item{object, object2, model}{An object of class \code{"ivreg"}.}

\item{...}{arguments to pass down.}

\item{x}{An object of class \code{"ivreg"} or \code{"summary.ivreg"}.}

\item{component}{For \code{terms}, \code{"regressors"} or \code{"instruments"}; 
for \code{model.matrix}, \code{"projected"}, \code{"regressors"}, or \code{"instruments"}.}

\item{newdata}{Values of predictors for which to obtain predicted values.}

\item{na.action}{\code{na} method to apply to predictor values for predictions; default is \code{\link{na.pass}}.}

\item{digits}{For printing.}

\item{vcov.}{Optional coefficient covariance matrix to use in computing model summary.}

\item{df}{Optional residual degrees of freedom to use in computing model summary.}

\item{diagnostics}{Report 2SLS diagnostic tests in model summary (default is \code{FALSE}).}

\item{signif.stars}{Show "significance stars" in summary output.}

\item{test}{Test statistics for ANOVA table; only \code{test = "F"} is supported.}

\item{formula.}{To update model.}

\item{evaluate}{If \code{TRUE}, the default, the updated model is evaluated; if \code{FALSE} the updated call is returned.}
}
\description{
Various methods for processing \code{"ivreg"} objects; for diagnostic methods,
  see ivregDiagnostics.
}
