% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivreg.R
\name{model.matrix.ivreg}
\alias{model.matrix.ivreg}
\alias{2SLS_Methods}
\alias{avPlot.ivreg}
\alias{vcov.ivreg}
\alias{residuals.ivreg}
\alias{fitted.ivreg}
\alias{print.ivreg}
\alias{summary.ivreg}
\alias{print.summary.ivreg}
\alias{anova.ivreg}
\alias{update.ivreg}
\alias{bread.ivreg}
\alias{estfun.ivreg}
\title{Methods for \code{"ivreg"} Objects}
\usage{
\method{model.matrix}{ivreg}(object, type = c("model", "instruments",
  "stage2"), ...)

\method{avPlot}{ivreg}(model, ...)

\method{vcov}{ivreg}(object, ...)

\method{residuals}{ivreg}(object, type = c("response", "stage1",
  "stage2", "working", "deviance", "pearson", "partial"), ...)

\method{fitted}{ivreg}(object, type = c("model", "stage1", "stage2"),
  ...)

\method{print}{ivreg}(x, digits = getOption("digits") - 2, ...)

\method{summary}{ivreg}(object, digits = getOption("digits") - 2,
  vcov. = vcov, ...)

\method{print}{summary.ivreg}(x, ...)

\method{anova}{ivreg}(object, model.2, s2, dfe, ...)

\method{update}{ivreg}(object, formula., instruments., ...,
  evaluate = TRUE)

\method{bread}{ivreg}(x, ...)

\method{estfun}{ivreg}(x, ...)
}
\arguments{
\item{object}{An object of class \code{"ivreg"}.}

\item{type}{Type of object desired, varies by method:
\describe{
  \item{\code{model.matrix}}{\code{"model"} (the default), \code{"instruments"}, or
    \code{"stage2"}.}
  \item{\code{residuals}}{\code{"response"}, \code{"stage1"}, \code{"stage2"},
    \code{"working"} (equivalent to \code{"response"}), \code{"deviance"},
    \code{"pearson"} (equivalent to \code{"deviance"}), or \code{"partial"}.}
  \item{\code{fitted}}{\code{"model"}, \code{"stage1"}, or \code{"stage2"}.}
  }}

\item{...}{to match generics, not generally used.}

\item{model}{An object of class \code{"ivreg"} or \code{"influence.ivreg"}.}

\item{x}{An object of class \code{"ivreg"}.}

\item{digits}{Digits to print.}

\item{vcov.}{Function to compute the coefficient covariance matrix or the matrix itself;
the default is the function \code{vcov}; set, e.g., to \code{\link[sandwich]{sandwich}}
(from the \pkg{sandwich} package) to get robust coefficient standard errors.}

\item{model.2}{A second object of class \code{"ivreg"}.}

\item{s2}{the estimated error variance (optional);
if not specified, taken from the larger model.}

\item{dfe}{the degrees of freedom for error (optional);
if not specified, taken from the larger model.}

\item{formula.}{Updated model formula.}

\item{instruments.}{Updated one-sided formula for the instrumental variables.}

\item{evaluate}{If \code{TRUE} (the default) evaluate the updated model;
if \code{FALSE} simply generate the updated call.}
}
\description{
Various methods for processing \code{"ivreg"} objects; for diagnostic methods,
  see \link{2SLS_Diagnostics}.
}
\examples{
kmenta.eq1 <- ivreg(Q ~ P + D, ~ D + F + A, data=Kmenta)
coef(kmenta.eq1) # estimates
sqrt(diag(vcov(kmenta.eq1))) # std. errors
summary(kmenta.eq1)
summary(kmenta.eq1, vcov.=sandwich::sandwich) # sandwich SEs
plot(fitted(kmenta.eq1), residuals(kmenta.eq1)) # residuals vs fitted values
}
\seealso{
\code{\link{ivreg}}, \link{2SLS_Diagnostics},
  \code{\link[sandwich]{sandwich}}
}
