% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivreg2.R
\name{Rsq}
\alias{Rsq}
\alias{Rsq.default}
\title{R-Squares Analog}
\usage{
Rsq(model, ...)

\method{Rsq}{default}(model, adjusted = FALSE, ...)
}
\arguments{
\item{model}{A model object.}

\item{...}{Possible arguments for specific methods.}

\item{adjusted}{If \code{TRUE} (the default is \code{FALSE}) return the \eqn{R^2} adjusted
for degrees of freedom.}
}
\description{
R-Squares Analog
}
\examples{
Rsq(ivreg2(Q ~ P + D, ~ D + F + A, data=Kmenta))
}
